+++ packages/perseus/src/index.ts
@@ -89,6 +89,10 @@ export {
     genericPerseusItemData,
 } from "./util/test-utils";
 export {
+    convertWidgetNameToEnum,
+    getWidgetTypeFromWidgetId,
+} from "./util/widget-enum-utils";
+export {
     WidgetType,
     addWidget,
     getAllWidgetIds,